Minor cleanup - made seperate category specific helper
